Update README and add RAG sample (#81)
* Update README

* Add acme-corp-rag sample app

* Remove basic retrieval samples

* Move sample to langchain dir

* Add pebblo logo

---------
