+++ samples/langchain/acme-corp-rag/.env
@@ -1,0 +1,2 @@
+OPENAI_API_KEY=""
+PEBBLO_CLASSIFIER_URL="http://localhost:8000/"
