+++ samples/langchain/acme-corp-rag/acme_corp_rag.py
@@ -1,27 +1,40 @@
-from typing import Any, List, Optional, Sequence
-import logging
-from dotenv import load_dotenv
-load_dotenv()
-
 from langchain.chains import RetrievalQA
-from langchain_community.document_loaders import DaxaSafeLoader
 from langchain.document_loaders.csv_loader import CSVLoader
 from langchain_openai.embeddings import OpenAIEmbeddings
 from langchain_openai.llms import OpenAI
 from langchain.schema import Document
-from langchain.schema.output import LLMResult
 from langchain_community.vectorstores import Chroma
-from langchain.vectorstores.utils import filter_complex_metadata
+from langchain_community.vectorstores.utils import filter_complex_metadata
+from typing import List
 
-logging.basicConfig(level=10)
-class OpenAIGenieCsv:
+
+# Fill-in OPENAI_API_KEY in .env file
+# in this directory before proceeding
+
+from dotenv import load_dotenv
+load_dotenv()
+
+class AcmeCorpRAG:
     def __init__(self, file_path: str):
-        self.loader = DaxaSafeLoader(
-            CSVLoader(file_path), "csv_app_1", "rahul"
-            )
+        self.file_path = file_path
+        self.app_name = "acme-corp-rag-1"
+
+        # Load documents
+
+        print(f"Loading RAG documents ...")
+        self.loader = CSVLoader(self.file_path)
         self.documents = self.loader.load()
         self.filtered_docs = filter_complex_metadata(self.documents)
+        print(f"Loaded {len(self.documents)} documents ...\n")
+
+        # Load documents into VectorDB
+
+        print(f"Hydrating Vector DB ...")
         self.vectordb = self.embeddings(self.filtered_docs)
+        print(f"Finished hydrating Vector DB ...\n")
+
+        # Prepare retriever QA chain
+
         llm = OpenAI()
         self.retriever = RetrievalQA.from_chain_type(
             llm=llm,
@@ -37,12 +50,12 @@ class OpenAIGenieCsv:
         return vectordb
 
     def ask(self, query: str):
-        return self.retriever.run(query)
+        return self.retriever.invoke(query)
 
 
 if __name__ == "__main__":
-    file_path = "../data/sens_data.csv"
-    genie = OpenAIGenieCsv(file_path)
-    prompt = "What does 213.85.121.199 mean?"
-    response = genie.ask(prompt)
+    rag_app = AcmeCorpRAG("./data/topic_data.csv")
+    prompt = "What is adaptive pacing system?"
+    print(f"Query:\n{prompt}")
+    response = rag_app.ask(prompt)
     print(f"Response:\n{response}")
