+++ samples/langchain/acme-corp-rag/data/topic_data.csv
@@ -1,0 +1,5 @@
+Content
+"Wipros board on Friday, January 12 announced an interim dividend of Re 1 per equity share of the face value of Rs 2 each, i.e., a 50 per cent payout for the current financial year along with financial results for the October-December period of the company for the financial year ending March 2024."
+"Roberts reminded the board of the scheduled retreat coming up in three months, and provided a drafted retreat schedule. The board provided feedback on the agenda and the consensus was that, outside of making a few minor changes, the committee should move forward as planned. No board action required."
+"Claims: An adaptive pacing system for implantable cardiac devices, comprising a pulse generator, multiple sensing electrodes, a microprocessor-based control unit, a wireless communication module, and memory for dynamically adjusting pacing parameters based on real-time physiological data.  The system of claim 1, wherein the adaptive pacing algorithms include rate-responsive pacing based on physical activity.  The system of claim 1, further comprising an external monitoring system for remote data access and modification of pacing parameters."
+"Sachin's SSN is 222-85-4836. His passport ID is 5484880UA.  Sachin's driver's license number is S9998888.  Sachin's bank account number is 70048841700216300.  His American express credit card number is 371449635398431.  His UK IBAN Code is AZ96AZEJ00000000001234567890.  ITIN number 993-77 0690. Azure client secret : c4cb6f91-15a7-4e6d-a824-abcdef012345.  AWS Access Key is: AKIAQIPT4PDORIRTV6PH AWS Secret Key is : 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 Github Token is: ghp_hgu657yiujgwfrtigu3ver238765tyuhygvtrder6t7gyvhbuy5e676578976tyghy76578uygfyfgcyturtdf Google API key: zaCELgL0imfnc8mVLWwsAawjYr4Rx-Af50DDqtlx Slack Token is: xoxp-7676545380258-uygh Azure Client Secret - c4cb6f91-15a7-4e6d-a824-abcdef012345 Slack Token - xoxb-3204014939555-4519358291237-TTIf0243T8YFSAGEVr1wBrWE Google API key- KLzaSyB_tWrbmfWx8g2bzL7Vhq7znuTUn0JPKmY"
